A standard deck of playing cards consists of 52 cards.
A standard deck of playing cards consists of 52 cards. All cards are divided
into 4 suits including spades (♠), clubs (♣), diamonds (♦) and hearts (♥). In
each suit there are 13 cards including a 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, J (jack), Q
(queen), K (king). A 5–card poker hand is said to be a “three of a kind” if it
consists of 3 cards of the same denomination, and 2 other cards of 2 different
denominations other than the one for the previous 3 cards.
(a) Making a list of all possible 5 card poker and assign to variable poker_5.
Assign number of 5 card poker to variable len_poker_5.
(b) Find all of “three of a kind” and assign to variable three_of_a_kind.
Assign the number to variable len_three_of_a_kind.